+++ RELEASENOTES.md
@@ -72,6 +72,9 @@
 *   Video:
     *   Add support for skipping frames that are late during join rather than
         dropping in DecoderVideoRenderer.
+    *   Adjust logic for codec reuse at frame rate changes on API<30 to avoid
+        codec resets where they are not beneficial
+        ([#3120](https://github.com/androidx/media/issues/3120)).
 *   Text:
 *   Metadata:
 *   Image: